test/client/requirements.txt
astroid==2.3.3
asttokens==1.1.13
attrdict==2.0.1
attrs==19.3.0
base58==1.0.3
certifi==2019.11.28
chardet==3.0.4
cytoolz==0.10.1
eth-abi==2.1.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
grpcio==1.25.0
grpcio-tools==1.25.0
hexbytes==0.2.0
idna==2.8
importlib-metadata==1.3.0
ipfshttpclient==0.4.12
isort==4.3.21
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lru-dict==1.1.6
mccabe==0.6.1
more-itertools==8.0.2
multiaddr==0.0.8
mypy-extensions==0.4.3
netaddr==0.7.19
parsimonious==0.8.1
pipenv==2018.11.26
protobuf==3.11.1
pycryptodome==3.9.4
pylint==2.4.4
pyrsistent==0.15.6
requests==2.22.0
rlp==1.2.0
six==1.13.0
toolz==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
varint==1.0.2
virtualenv==16.7.9
virtualenv-clone==0.5.3
vyper==0.1.0b14
web3==5.4.0
websockets==8.1
wrapt==1.11.2
zipp==0.6.0
